Generalize FrequencySlider as LogSlider, and add CenterValue parameter
The slider uses logarithmic-like scale.

However, CenterValue allows the user shift the values within the slider.
Min and Max doesn't move, but CenterValue will be the center of the slider.
CenterValue could be 440Hz, 1760Hz, volume 0.1/1...

The user can disable this feature to get a logarithmic scale by setting CenterValue to the value of Minimum.
Note that if Maximum is close to 1, then the scale is not logarithmic with this implementation. So it is recommended to set CenterValue.